Updated SL Ticket to 2014 layout (50%)
While I did simply add a comma and a newline after mTime (as of the new look) I didn't change the sCode (should be 6 integers then a space then the rest of the integers) because I'm not a Java programmer and your code confused me. I didn't want to screw anything up.